project/requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
dj-database-url==0.5.0
Django==2.2.2
django-redis==4.10.0
djangorestframework==3.9.4
kombu==4.6.1
mysqlclient==1.4.2
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
uWSGI==2.0.18
vine==1.3.0
